requirements_pip.txt
psycopg2-binary
pandas
python-bcb
findspark
yfinance
dbt-core
dbt-postgres